# qa-engineer/backend/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: To Do API
  license:
    name: MIT
servers:
  - url: "http://localhost:3001/"
  - url: "https://api.dev.cuvama.com/"
paths:
  /tasks:
    get:
      summary: Get the list of tasks
      operationId: getTasks
      tags:
        - tasks
      responses:
        "200":
          description: List of Tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tasks"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new task
      operationId: createTask
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "200":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}:
    delete:
      summary: Delete a task
      operationId: deleteTask
      tags:
        - tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: The version of the master data to retrieve
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateTask:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
    Error:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
    Task:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
    Tasks:
      type: array
      items:
        $ref: "#/components/schemas/Task"